+++ features/videoplayer/src/main/java/me/pitok/videoplayer/states/OptionsState.kt
@@ -1,0 +1,8 @@
+package me.pitok.videoplayer.states
+
+sealed class OptionsState: VideoPlayerState() {
+    object ShowMainMenu: OptionsState()
+    object ShowSubtitleMenu: OptionsState()
+    object ShowPlaybackSpeedMenu: OptionsState()
+    object ShowAudioMenu: OptionsState()
+}
