+++ features/videoplayer/src/main/java/me/pitok/videoplayer/states/SubtitleState.kt
@@ -1,0 +1,8 @@
+package me.pitok.videoplayer.states
+
+sealed class SubtitleState: VideoPlayerState() {
+    object Clear: SubtitleState()
+    class Show(val subText: String): SubtitleState()
+    object SubtitleNotFoundError: SubtitleState()
+    object SubtitleReadingError: SubtitleState()
+}
